Started styling nav bar
