+++ client/assets/scss/NavBar.scss
@@ -1,0 +1,13 @@
+.navStyle {
+  // border: 2px solid black;
+  background-color: lightblue;
+  color: darkslateblue;
+  height: 3rem;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  a {
+    font-size: large;
+    text-decoration: none;
+  }
+}
