+++ client/assets/scss/reset.scss
@@ -1,0 +1,27 @@
+/* imported by App.jsx */
+body,
+h1,
+h2,
+h3,
+h4,
+p,
+table,
+th,
+tr,
+td,
+ul,
+footer {
+  margin: 0px 0px 0px 0px;
+  padding: 0px 0px 0px 0px;
+  border: 0px 0px 0px 0px;
+  font-size: 100%;
+  font: inherit;
+}
+
+body {
+  min-height: 100vh;
+}
+
+.root {
+  min-height: 100%;
+}
