+++ client/assets/scss/test.scss
@@ -1,0 +1,3 @@
+// span {
+//     color: red;
+// }
